kosli-config-monitoring/requirements.txt
pytest==6.2.5
PyYAML
